add application module
